Support resolving authorizers by name
Changes:
 -  Move resolvers to a directory structure and break tests into parts
 -  Add a null authorizer binding
 -  Allow authorizers to be passed by name
