+++ src/authorizers/index.js
@@ -1,0 +1,6 @@
+import { bind } from '@globality/nodule-config';
+
+
+/* Register a default 'null' authorizer.
+ */
+bind('graphql.authorizers.null', () => () => true);
