+++ src/resolvers/__tests__/test.resolver.js
@@ -1,9 +1,5 @@
-import { Nodule } from '@globality/nodule-config';
-
 import {
     createResolver,
-    createStrictResolver,
-    getResolver,
     Forbidden,
 } from 'index';
 
@@ -73,63 +69,13 @@ describe('a resolver', () => {
         });
         await expect(resolver.resolve()).rejects.toThrow('Forbidden');
     });
-});
 
-
-describe('null resolver', () => {
-    beforeEach(async () => {
-        await Nodule.testing().load();
-    });
-
-    it('is available by name', async () => {
-        const resolver = getResolver('null');
-        expect(resolver).toBeDefined();
-
-        const result = await resolver();
-        expect(result).toBeNull();
-    });
-
-    it('is available by function call', async () => {
-        const resolver = getResolver(() => 'null');
-        expect(resolver).toBeDefined();
-
-        const result = await resolver();
-        expect(result).toBeNull();
-    });
-});
-
-
-describe('a strict resolver', () => {
-    it('requires an aggregate function', async () => {
-        expect(
-            () => createStrictResolver({
-                authorize: () => true,
-                transform: value => value,
-            }),
-        ).toThrow(
-            'Strict resolver must define an `aggregate` option',
-        );
-    });
-
-    it('requires an authorize function', async () => {
-        expect(
-            () => createStrictResolver({
-                aggregate: async () => null,
-                transform: value => value,
-            }),
-        ).toThrow(
-            'Strict resolver must define an `authorize` option',
-        );
-    });
-
-    it('requires a transform function', async () => {
-        expect(
-            () => createStrictResolver({
-                aggregate: async () => null,
-                authorize: () => true,
-            }),
-        ).toThrow(
-            'Strict resolver must define a `transform` option',
-        );
+    it('authorizes based on an authorizer name', async () => {
+        const resolver = createResolver({
+            aggregate: async () => 21,
+            authorize: 'null',
+            transform: result => 2 * result,
+        });
+        expect(await resolver.resolve()).toEqual(42);
     });
 });
